feat: 모듈화 2차
