+++ src/components/three/NightSky/NightSky.tsx
@@ -3,6 +3,7 @@ import { useEffect, useRef } from 'react'
 import * as THREE from 'three'
 import { createThreeScene } from './threeSetup'
 import { createTitleElement, createButtonElement } from './createUIElements'
+import { createTorus, createParticles } from './createObjects'
 import { handleWindowResize } from './eventHandlers'
 
 interface NightSkyProps {
@@ -37,50 +38,11 @@ const NightSky: React.FC<NightSkyProps> = () => {
 
     // Objects
     // Sphere
-    const geometry = new THREE.TorusGeometry(0.2, 0.03, 54, 100)
-    const material = new THREE.MeshPhysicalMaterial({
-      color: 0x78e6e6,
-      metalness: 2,
-      roughness: 0.9,
-      clearcoat: 1,
-      clearcoatRoughness: 1,
-      side: THREE.DoubleSide,
-      envMapIntensity: 5,
-      emissive: 0xa4aa0,
-      emissiveIntensity: 10,
-    })
-    const sphere = new THREE.Mesh(geometry, material)
+    const sphere = createTorus()
     scene.add(sphere)
 
     // Particle
-    const particlesGeometry = new THREE.BufferGeometry()
-    const loader = new THREE.TextureLoader()
-    const canvas = document.createElement('canvas')
-    const ctx = canvas.getContext('2d')!
-    canvas.height = 100
-    canvas.width = 100
-    ctx.fillStyle = '#fff'
-    ctx.beginPath()
-    ctx.arc(10, 50, 25, 0, 2 * Math.PI)
-    ctx.fill()
-    const img = canvas.toDataURL('image/png')
-    const star = loader.load(img)
-    const particlesmaterial = new THREE.PointsMaterial({
-      size: 0.01,
-      map: star,
-      transparent: true,
-    })
-    const particlesCnt = 5500
-    const posArray = new Float32Array(particlesCnt * 3)
-    const colorsArray = new Float32Array(particlesCnt * 3)
-    for (let i = 0; i < particlesCnt * 3; i++) {
-      posArray[i] = (Math.random() - 0.5) * (Math.random() * 4)
-      colorsArray[i] = Math.random()
-    }
-
-    particlesGeometry.setAttribute('position', new THREE.BufferAttribute(posArray, 3))
-    particlesGeometry.setAttribute('color', new THREE.BufferAttribute(colorsArray, 3))
-    const particlesMesh = new THREE.Points(particlesGeometry, particlesmaterial)
+    const { mesh: particlesMesh, geometry: particlesGeometry, particlesCnt } = createParticles()
     scene.add(particlesMesh)
 
     // Mouse variables
